Move all SwiftUI classes to RealmSwift/SwiftUI.swift
