+++ RealmSwift/Object.swift
@@ -846,88 +846,3 @@ extension Object: AssistedObjectiveCBridgeable {
         return (objectiveCValue: unsafeCastToRLMObject(), metadata: nil)
     }
 }
-
-#if canImport(SwiftUI)
-
-import SwiftUI
-@dynamicMemberLookup
-@available(iOS 13.0, macOS 10.15, tvOS 13.0, watchOS 6.0, *)
-public class BoundObject<T: Object> : ObservableObject {
-    internal var value: T
-    init(_ value: T) {
-        self.value = value
-    }
-
-    public subscript<V>(dynamicMember member: ReferenceWritableKeyPath<T, V>) -> Binding<V> {
-        get {
-            Binding(get: {
-                self.value[keyPath: member]
-            }, set: { newValue in
-                try! self.value.realm!.write {
-                    self.value[keyPath: member] = newValue
-                }
-            })
-        }
-    }
-
-    // if property is a list, we want to freeze it before fetching it
-    public subscript<V>(dynamicMember member: ReferenceWritableKeyPath<T, V>) -> Binding<V> where V: ListBase {
-        get {
-            Binding(get: {
-                self.value[keyPath: member].freeze()
-            }, set: {
-                self.value[keyPath: member] = $0
-            })
-        }
-    }
-}
-
-@available(iOS 13.0, macOS 10.15, tvOS 13.0, watchOS 6.0, *)
-public extension Binding where Value: Object {
-    subscript<V>(dynamicMember member: ReferenceWritableKeyPath<Value, V>) -> Binding<V> {
-        get {
-            Binding<V>(get: {
-                self.wrappedValue[keyPath: member]
-            }, set: { newValue in
-                try! wrappedValue.realm!.write {
-                    self.wrappedValue[keyPath: member] = newValue
-                }
-            })
-        }
-    }
-
-    subscript<V>(dynamicMember member: ReferenceWritableKeyPath<Value, V>) -> Binding<V> where V: ListBase {
-        get {
-            Binding<V>(get: {
-                self.wrappedValue[keyPath: member].freeze()
-            }, set: {
-                self.wrappedValue[keyPath: member] = $0
-            })
-        }
-    }
-}
-
-@available(iOS 13.0, macOS 10.15, tvOS 13.0, watchOS 6.0, *)
-@propertyWrapper
-public struct BindObject<T: Object>: DynamicProperty {
-    private var _wrappedValue: ObservedObject<BoundObject<T>>
-    public var wrappedValue: T {
-        get {
-            _wrappedValue.wrappedValue.value
-        }
-    }
-    /// The property that can be accessed with the `$` syntax and allows access to the `Publisher`
-    public var projectedValue: BoundObject<T> {
-        return _wrappedValue.wrappedValue
-    }
-    var token: NotificationToken?
-
-    public init(wrappedValue: T) {
-        _wrappedValue = ObservedObject(wrappedValue: BoundObject(wrappedValue))
-        token = wrappedValue.observe { [_wrappedValue] _ in
-            _wrappedValue.wrappedValue.objectWillChange.send()
-        }
-    }
-}
-#endif
-
