+++ RealmSwift/Results.swift
@@ -452,54 +452,3 @@ extension Results: Encodable where Element: Encodable {
         }
     }
 }
-
-#if canImport(SwiftUI)
-import SwiftUI
-
-@available(OSX 10.15, watchOS 6.0, iOS 13.0, iOSApplicationExtension 13.0, OSXApplicationExtension 10.15, tvOS 13.0, *)
-public final class BoundResults<Element: RealmCollectionValue>: ResultsBase, ObservableObject {
-    public typealias Element = Element
-    let rlmResults: RLMResults<AnyObject>
-
-    init(_ results: Results<Element>) {
-        self.rlmResults = results.rlmResults
-    }
-    init(_ rlmResults: RLMResults<AnyObject>) {
-        self.rlmResults = rlmResults
-    }
-    init(objc: RLMResults<AnyObject>) {
-        self.rlmResults = objc
-    }
-}
-
-@available(OSX 10.15, watchOS 6.0, iOS 13.0, iOSApplicationExtension 13.0, OSXApplicationExtension 10.15, tvOS 13.0, *)
-@propertyWrapper
-public struct BindResults<T>: DynamicProperty where T: Object {
-    private var _wrappedValue: ObservedObject<BoundResults<T>>
-    public var wrappedValue: BoundResults<T> {
-        get {
-            _wrappedValue.wrappedValue
-        }
-    }
-    var token: NotificationToken?
-
-    public init(_ type: T.Type, realm: Realm? = nil) {
-        if let realm = realm {
-            _wrappedValue = ObservedObject(wrappedValue: BoundResults(realm.objects(T.self)))
-        } else {
-            _wrappedValue = ObservedObject(wrappedValue: BoundResults(Environment(\.realm).wrappedValue.objects(T.self)))
-        }
-        token = self._wrappedValue.wrappedValue.observe { [_wrappedValue] value in
-            switch value {
-            case .initial(_):
-                _wrappedValue.wrappedValue.objectWillChange.send()
-            case .update(_, _, _, _):
-                _wrappedValue.wrappedValue.objectWillChange.send()
-            case .error(_):
-                fatalError()
-            }
-        }
-    }
-}
-
-#endif
