--- examples/ios/swift/ListSwiftUI/SceneDelegate.swift
@@ -1,44 +1,0 @@
-////////////////////////////////////////////////////////////////////////////
-//
-// Copyright 2020 Realm Inc.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-// http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-//
-////////////////////////////////////////////////////////////////////////////
-
-import UIKit
-import SwiftUI
-import RealmSwift
-
-class SceneDelegate: UIResponder, UIWindowSceneDelegate {
-    var window: UIWindow?
-
-    func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
-        guard let windowScene = scene as? UIWindowScene else {
-            return
-        }
-
-        let window = UIWindow(windowScene: windowScene)
-
-        let realm = try! Realm()
-        var recipes = realm.object(ofType: Recipes.self, forPrimaryKey: 0)
-        if recipes == nil {
-            recipes = try! realm.write { realm.create(Recipes.self, value: []) }
-        }
-
-        window.rootViewController = UIHostingController(rootView: ContentView(recipes: recipes!.recipes))
-
-        self.window = window
-        window.makeKeyAndVisible()
-    }
-}
