--- examples/ios/swift/ListSwiftUI/Views/SearchBar.swift
@@ -1,34 +1,0 @@
-////////////////////////////////////////////////////////////////////////////
-//
-// Copyright 2020 Realm Inc.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-// http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-//
-////////////////////////////////////////////////////////////////////////////
-
-import SwiftUI
-
-struct SearchBar: View {
-    @Binding var text: String
-
-    var body: some View {
-        ZStack {
-            HStack {
-                Image(systemName: "magnifyingglass")
-                TextField("recipe or ingredient", text: $text).scaledToFill()
-            }.padding()
-            RoundedRectangle(cornerRadius: 10)
-                .stroke(Color(red: 0.9, green: 0.9, blue: 0.9), lineWidth: 3)
-        }
-    }
-}
